Initial Commit
Working Login, Barrack, and getting In-game using the latest client.

- In game bugged, nothing works yet, you can walk around.

+++ src/Shared/Network/Helpers/CommanderHelper.cs
@@ -35,12 +35,12 @@ namespace Melia.Shared.Network.Helpers
 
 			packet.PutInt(commander.Hp);
 			packet.PutInt(commander.MaxHp);
-			packet.PutShort(commander.Sp);
-			packet.PutShort(commander.MaxSp);
-			packet.PutInt(commander.Stamina);
-			packet.PutInt(commander.MaxStamina);
-			packet.PutShort(0); // Shield
-			packet.PutShort(0); // MaxShield
+			packet.PutInt(commander.Sp);
+			packet.PutInt(commander.MaxSp);
+			packet.PutInt(commander.Stamina); // Not stamina (Value: 1000)
+			packet.PutInt(commander.MaxStamina); // Not max stamina (Value: 25000)
+			packet.PutInt(0); // Shield
+			packet.PutInt(0); // MaxShield
 		}
 	}
